Added a new field Category for the Catalog section. +Removed the unused MySession object. +Now the search in the index page can only take place if the search textbox is not empty.

+++ app/models/form/FormCategory.scala
@@ -2,5 +2,5 @@ package models.form
 
 import models.OBCategory
 
-case class FormCategory(selectedID: Option[Long], updatedCategoryName: String, 
+case class FormCategory(selectedID: Option[Int], updatedCategoryName: String, 
     newCategoryName: String, list: Option[List[OBCategory]])
